SIS4/requirements.txt
streamlit==1.32.2
pytesseract==0.3.10
pdf2image==1.17.0
openai==1.14.0
python-dotenv==1.0.1
Pillow==10.2.0
pydantic==2.6.4
